Updated version of the mcast rate adaptation
